list-faq refactor
