+++ src/components/VueDrawerHeader.vue
@@ -3,11 +3,13 @@ import { defineComponent, ref, reactive, computed } from 'vue';
 
 // import components
 import VueMenuLinks from "./VueMenuLinks.vue"
+import VueListFaq from './VueListFaq.vue';
 
 export default defineComponent({
   name: 'VueDrawerHeader',
   components: {
-    VueMenuLinks
+    VueMenuLinks,
+    VueListFaq,
   },
   props: {
     showLogo: {
@@ -27,88 +29,6 @@ export default defineComponent({
   setup(props, { emit }) {
     const dialogOpened = ref(false);
     const dialogState = ref('default');
-
-    const dummyText =
-      'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.';
-    const itemsFAQParticipant = [
-      {
-        title: 'Chci z toho vycouvat. Vrátite mi startovné?',
-        text: dummyText,
-      },
-      {
-        title: 'Nechcete přijmout moji platbu startovného.',
-        text: dummyText,
-      },
-      {
-        title:
-          'Do práce mi to na kole deska nějak nevyšlo. Můžu si místo toho zapsat víkendový vylet?',
-        text: dummyText,
-      },
-      {
-        title: 'Jak se propojím se svojí mobilní aplikací?',
-        text: dummyText,
-      },
-      {
-        title: 'Jak si zapíšu cestu na noční směnu?',
-        text: dummyText,
-      },
-      {
-        title: 'Můžu na cestě do práce kombinovat dopravní prostředky?',
-        text: dummyText,
-      },
-      {
-        title: 'Chci změnit adresu, na kterou mi pošlete startovní balíček.',
-        text: dummyText,
-      },
-      {
-        title: 'Ještě nemám startovní balíček, safra.',
-        text: dummyText,
-      },
-      {
-        title: 'Nechci to moc rozebírat, ale velikost trička mi nesedí.',
-        text: dummyText,
-      },
-    ];
-    const itemsFAQCoordinator = [
-      {
-        title: 'Jsem hodný šéf a chci zaplatit za své zaměstnance.',
-        text: dummyText,
-      },
-      {
-        title: 'Chci predat břímě firemniho koordinátora někomu jinému.',
-        text: dummyText,
-      },
-      {
-        title: 'Chci vidět výsledky výzev.',
-        text: dummyText,
-      },
-      {
-        title: 'Jak se stanu hrdým firemním koordinátorem?',
-        text: dummyText,
-      },
-      {
-        title:
-          'Jak se stát firemním koordinátorem a zároveň se účastnit výzvy Do práce na kole?',
-        text: dummyText,
-      },
-      {
-        title:
-          'Nechcete se účastnit výzvy, ale chcete se stát firemním koordinátorem?',
-        text: dummyText,
-      },
-      {
-        title: 'Jak uhradit startovné za kolegy?',
-        text: dummyText,
-      },
-      {
-        title: 'Jak se hlásí kolegové?',
-        text: dummyText,
-      },
-      {
-        title: 'Kde najdu fakturu?',
-        text: dummyText,
-      },
-    ];
 
     const classes = computed(() => {
       return props.showLogo ? 'justify-between' : 'justify-end';
@@ -133,8 +53,6 @@ export default defineComponent({
       drawerOpened,
       dialogOpened,
       dialogState,
-      itemsFAQParticipant,
-      itemsFAQCoordinator,
       contactForm,
     };
   },
@@ -220,45 +138,8 @@ export default defineComponent({
           data-cy="dialog-content"
           style="max-height: 50vh"
         >
-          <h4
-            class="text-h5 text-weight-bold q-my-none q-px-md"
-            data-cy="title-participants"
-          >
-            {{ $t('index.help.titleParticipants') }}
-          </h4>
-          <q-list separator class="faq-list q-mt-md" data-cy="faq-participants">
-            <q-expansion-item
-              v-for="item in itemsFAQParticipant"
-              :key="item.title"
-              :label="item.title"
-            >
-              <q-card>
-                <q-card-section>
-                  {{ item.text }}
-                </q-card-section>
-              </q-card>
-            </q-expansion-item>
-          </q-list>
-
-          <h4
-            class="text-h5 text-weight-bold q-my-none q-px-md q-mt-xl"
-            data-cy="title-coordinators"
-          >
-            {{ $t('index.help.titleCoordinators') }}
-          </h4>
-          <q-list separator class="faq-list q-mt-md" data-cy="faq-coordinators">
-            <q-expansion-item
-              v-for="item in itemsFAQCoordinator"
-              :key="item.title"
-              :label="item.title"
-            >
-              <q-card>
-                <q-card-section>
-                  {{ item.text }}
-                </q-card-section>
-              </q-card>
-            </q-expansion-item>
-          </q-list>
+          <vue-list-faq :title="$t('index.help.titleParticipants')" variant="participant"></vue-list-faq>
+          <vue-list-faq :title="$t('index.help.titleCoordinators')" variant="coordinator" class="q-mt-xl"></vue-list-faq>
 
           <div class="q-px-md q-mt-xl">
             <h4
